+++ CHANGELOG.md
@@ -11,16 +11,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 
+- added managed autoscaling config to `sagemaker-endpoint` module
 - added SSO support in `sagemaker-studio` module
 - added VPC/subnets/sg config for multi-account project template to `sagemaker-templates-service-catalog` module
 - added `sagemaker-custom-kernel` module
 - added batch inference project template to `sagemaker-templates-service-catalog` module
 - added EFS removal policy to `mlflow-fargate` module
 - added `mwaa` module with example dag which demonstrates the MLOps in Airflow
+- added `sagemaker-model-event-bus` module.
+- added `sagemaker-model-package-group` module.
+- added `sagemaker-model-package-promote-pipeline` module.
 - added `sagemaker-hugging-face-endpoint` module
+- added `hf_import_models` template to import hugging face models
 
 ### **Changed**
 
+- reogranized manifests by use-case
 - add account/region props for project templates in `sagemaker-templates-service-catalog` module
 - fix `sagemaker-templates-service-catalog` model deploy role lookup issue & abalone_xgboost model registry permissions
 - update `sagemaker-custom-kernel` module IAM permissions